Create folder per identity
